+++ test/new_eos_keys.test.mjs
@@ -1,9 +1,9 @@
 import { deepStrictEqual, rejects } from 'assert'
-import new_eos_keys from '../public/new_eos_keys.js'
-import public_key_from_private from '../public/public_key_from_private.js'
+import new_eos_keys from '../new_eos_keys.js'
+import public_key_from_private from '../public_key_from_private.js'
 
 export default tests => {
-  tests.add('new eos keys', async () => {
+  tests.add('recover public key from signature', async () => {
     const { public_key, private_key } = await new_eos_keys()
     deepStrictEqual(await public_key_from_private(private_key), public_key)